addresses and reg + model binding

+++ app/Rules/PhoneOccupied.php
@@ -19,14 +19,18 @@ class PhoneOccupied implements Rule
      */
     public function passes($attribute, $value)
     {
-        $same_model = Sms::where('phone','=', $value);
-        $same = $same_model->get()->first();
-        $now = Carbon::now();
-        $difference = $same->created_at->diffInSeconds($now);
-        if ($difference > 300) {
-            $same_model->forceDelete();
+        if (Sms::where('phone','=', $value)->exists()) {
+            $same_model = Sms::where('phone','=', $value);
+            $same = $same_model->get()->first();
+            $now = Carbon::now();
+            $difference = $same->created_at->diffInSeconds($now);
+            if ($difference > 300) {
+                $same_model->forceDelete();
+            }
+            return !User::where('phone', '=', $value)->exists() && !$same->where('confirmed', '=', 1)->exists() && $difference > 600;
+        } else {
+            return !User::where('phone', '=', $value)->exists() && !Sms::where('phone','=', $value)->where('confirmed', '=', 1)->exists();
         }
-        return !User::where('phone', '=', $value)->exists() && !$same->where('confirmed', '=', 1)->exists() && $difference > 600;
     }
 
     /**
